requirements.txt
Django==1.10.4
click==6.7
coverage==4.2
coveralls==1.1
django-rq==0.9.5
django-suit==0.2.23
django-tastypie==0.13.3
docopt==0.6.2
meld3==1.0.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
redis==2.10.5
requests==2.12.4
rq==0.7.1
six==1.10.0
supervisor==3.3.1
uWSGI==2.0.14
wsgiref==0.1.2
